pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools-scm[toml]>=6.2", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
write_to = "gensh/_version.py"

[tool.poetry]
name = "gensh"
version = "0.1.0"
description = "GenSh - Generate and execute code using natural language"
authors = ["Vaibhav Bhandari <vaibhavb@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
openai = "^1.0.0"
anthropic = "^0.3.0"
PyYAML = "^6.0"
ollama = "^0.1.0"
python-dotenv = "^0.19.0"
markdown = "^3.3.0"
python-docx = "^0.8.10"
requests = "^2.25.0"
beautifulsoup4 = "^4.9.3"

[tool.poetry.group.dev.dependencies]
pytest = "^6.2"

[tool.poetry.scripts]
gensh = "gensh.processor:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
